requirements.txt
streamlit
easyocr
numpy
pdfplumber
torch
inference-sdk
Pillow
pyzbar
